Challenge 1: Zombie Game! Pt2. Player is able to move now. He is invisible but able to move. Hopefully the zombies will never find him now.

+++ ZombieGame/MainGame.h
@@ -57,7 +57,7 @@ private:
 
 	int _screenWidth, _screenHeight;
 
-	int _fps;
+	float _fps;
 
 	int _currentLevel;
